move point-cloud transcoding out of draco.rs into the remote-access layer
draco.rs is the pure Draco encoding mechanism; the transcode submodule was
upper-layer sink glue -- it knew about RawChannel, channel schemas, and the
advertise wire protocol, and rewrote PointCloud advertisements to the
CompressedPointCloud schema. That belongs with the sink, not the encoder.

Relocate it to remote_access/point_cloud_transcode.rs (a sibling of the
point_cloud_compression policy module), repoint the three consumers
(channel_registry, point_cloud_compression, session/point_cloud_track) at the
new path, and drop the now-redundant nested `mod transcode` wrapper. The tests
move with it unchanged. draco.rs is left with just the encoding mechanism.
