+++ rust/foxglove/src/remote_access/session/point_cloud_track.rs
@@ -8,7 +8,7 @@ use tracing::warn;
 
 use crate::ChannelId;
 use crate::draco::CompressPointCloudOptions;
-use crate::draco::transcode::transcode_point_cloud_message;
+use crate::remote_access::point_cloud_transcode::transcode_point_cloud_message;
 
 use super::RemoteAccessSession;
